==> 1. Get_Company_List.ipynb <==
import pytest
import time
import json
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.support import expected_conditions
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
import pandas as pd

Year = [2021,2020]

data_final = pd.DataFrame()

for a in Year:
    driver = webdriver.Chrome()
    driver.get("https://www.idx.co.id/id/perusahaan-tercatat/laporan-keuangan-dan-tahunan")
    driver.set_window_size(1936, 1048)
    
    elements = driver.find_elements(By.CSS_SELECTOR, ".bzg_c:nth-child(4) > .mb-8 *")

    for element in elements:
        if element.text == "Tahunan":
            element.click()
    
    elements = driver.find_elements(By.CSS_SELECTOR, ".bzg_c:nth-child(3) > .mb-8 *")
    count = 0

    for element in elements:
        if element.text == str(a):
            element.click()

    driver.find_element(By.CSS_SELECTOR, ".btn--wider:nth-child(2)").click()

    element = driver.find_element(By.CSS_SELECTOR, ".btn--wider:nth-child(2)")
    actions = ActionChains(driver)
    actions.move_to_element(element).perform() 

    element = driver.find_element(By.CSS_SELECTOR, "body")
    actions = ActionChains(driver)

    data = pd.DataFrame()
    data2 = {}
    c = 0
    looping = True
    page = 1

    max_page = driver.find_element(By.CSS_SELECTOR, f".ph-8").text
    max_page = max_page.split(" ")
    max_page = max_page[1]  
    max_page = int(max_page)

    while looping:
        WebDriverWait(driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".bzg_c:nth-child(1) .box-title")))
        for i in range(1, 13):
            try:
                driver.find_element(By.CSS_SELECTOR, f".bzg_c:nth-child({i}) .box-title").click()
                companys = driver.find_element(By.CSS_SELECTOR, f".bzg_c:nth-child({i}) .f-20").text
                data2['Company'] = companys

                driver.find_element(By.CSS_SELECTOR, f".bzg_c:nth-child({i}) small").click()
                date = driver.find_element(By.CSS_SELECTOR, f".bzg_c:nth-child({i}) small").text
                data2['Date']= date

                driver.find_element(By.CSS_SELECTOR, f".bzg_c:nth-child({i}) tr:nth-child(2) > .full-width").click()
                Tahun_Buku = driver.find_element(By.CSS_SELECTOR, f".bzg_c:nth-child({i}) tr:nth-child(2) > .full-width").text
                data2['Tahun_Buku'] = Tahun_Buku
                data2 = pd.DataFrame(data2, index=[c])
                data = pd.concat([data, data2])

            except Exception as e:
                print(f"there in no company again")

        if page == max_page:
            looping = False
        else:
            try:
                driver.find_element(By.CSS_SELECTOR, ".\\--next").click()
            except:
                looping = False
        
        page += 1
        
    data.reset_index(drop=True, inplace=True)
    data.to_csv(f"Data_{a}.csv")
