==> 2. combine_list.ipynb <==
import os
import pandas as pd

current_path = os.getcwd()
raw_list_path = os.path.join(current_path, 'Raw List')
raw_list = os.listdir(raw_list_path)

df_combined = pd.DataFrame()
for i in raw_list:
    if i.endswith('.csv'):
        df = pd.read_csv(os.path.join(raw_list_path, i))
        df_combined = pd.concat([df_combined, df], axis=0)
    

df_combined.reset_index(drop=True, inplace=True)

df_combined

df_combined = df_combined.drop(['Unnamed: 0'], axis=1)

df_combined

def get_link(x):
    hasil = f'https://www.idx.co.id/Portals/0/StaticData/ListedCompanies/Corporate_Actions/New_Info_JSX/Jenis_Informasi/01_Laporan_Keuangan/02_Soft_Copy_Laporan_Keuangan//Laporan%20Keuangan%20Tahun%20{x["Tahun_Buku"]}/Audit/{x["Company"]}/inlineXBRL.zip'
    instancexbrl = f'https://www.idx.co.id/Portals/0/StaticData/ListedCompanies/Corporate_Actions/New_Info_JSX/Jenis_Informasi/01_Laporan_Keuangan/02_Soft_Copy_Laporan_Keuangan//Laporan%20Keuangan%20Tahun%20{x["Tahun_Buku"]}/Audit/{x["Company"]}/instance.zip'
    return instancexbrl

df_combined['Link'] = df_combined.apply(lambda x: get_link(x), axis=1)

df_combined

df_combined.to_csv('combine list.csv')
