==> 3. File_Download.ipynb <==
import pandas as pd
import os
import wget
import requests
import pytest
import time
import json
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.support import expected_conditions
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
import shutil
import time

current_dir = os.getcwd()

combine_list = pd.read_csv('combine list.csv')

co

combine_list

data_loss = pd.DataFrame()

for index,row in combine_list.iterrows():
    
    directory = os.path.join(current_dir,"Data", row['Company'], str(row['Tahun_Buku']))
    
    if not os.path.exists(directory):
        os.makedirs(directory)
    
    URL = str(row['Link'])
    download_file = os.path.join(directory, "instance.zip")

    driver = webdriver.Chrome()
    try:
        convert_path = r"C:\Users\arsen\Downloads"
        convert_path = os.path.normpath(convert_path)
        driver.get(URL)
        wait = WebDriverWait(driver, 60)
        wait.until(lambda driver: any("instance" in filename for filename in os.listdir(convert_path)))
        file_downlaod = os.path.join(convert_path, "instance.zip")
        shutil.move(file_downlaod, download_file)
    except:
        data={}
        data['Company'] = row['Company']
        data['Tahun_Buku'] = row['Tahun_Buku']
        data['Date'] = row['Date']
        data['Link'] = row['Link']
        data_loss = pd.concat([data_loss, pd.DataFrame([data])])
        
    driver.close()
data_loss.to_csv('data_loss.csv', index=False)
