==> 4. Unzip_Xbrl.ipynb <==
import pandas as pd
import zipfile
import os

combine_list = pd.read_csv('combine list.csv')

combine_list

curr_folder = os.getcwd()
data_folder = os.path.join(curr_folder, 'Data')

for index, row in combine_list.iterrows():
    file_directory =os.path.join(data_folder, row['Company'], str(row['Tahun_Buku']),'XBRL.zip')
    folder_directory = os.path.join(data_folder, row['Company'], str(row['Tahun_Buku']))
    try:
        if os.path.exists(file_directory):
            with zipfile.ZipFile(file_directory, 'r') as zip_ref:
                zip_ref.extractall(folder_directory)
    except:
        print('Error on ', row['Company'], row['Tahun_Buku'])
