==> 6a. Data_Process.ipynb <==
import pandas as pd
import yfinance as yf

# Load data_raw.csv

data = pd.read_csv('data_raw.csv')

data

# Changes Format Number to Float

header_exclude = ['Company','Tahun_Buku','Tanggal Rilis','category','End of Next Month']

header_list_changes =[]

for i in data.columns:
    if i not in header_exclude:
        header_list_changes.append(i)

header_list_changes

data[header_list_changes] = data[header_list_changes].replace({',': ''}, regex=True).astype(float)

def cogs(data):
    data_laba_bruto = data[data['laba_bruto'].isnull()]
    data_bruto = data[~data['laba_bruto'].isnull()]
    data_bruto['cogs'] = data_bruto['laba_bruto'] - data_bruto['pendapatan']
    data_laba_bruto['laba_bruto'] = data_laba_bruto['pendapatan'] + data_laba_bruto['cogs']
    data = pd.concat([data_bruto,data_laba_bruto])
    return data

def piutang_process(data):
    data_piutang = data[(data['piutang'].isnull()) | (data['piutang'] == 0)]
    data_piutang['piutang']=0.000001
    data_piutang_col = data[~((data['piutang'].isnull()) | (data['piutang'] == 0))]
    data = pd.concat([data_piutang,data_piutang_col])
    return data

data = data.drop(columns=['persediaan'])
data = cogs(data)
data['jumlah_aset_tidak_lancar'] = data['jumlah_aset'] - data['jumlah_aset_lancar']
data['jumlah_liabilitas_jangka_panjang'] = data['jumlah_liabilitas'] - data['jumlah_liabilitas_jangka_pendek']
data = data[data['7D_Avg'].notnull()]
data = data[data['pendapatan'].notnull()]
data = piutang_process(data)
#data = data.replace(0, 0.00001)

# Ratio Calculation

# Price Earning Ratio

data['per'] = data['7D_Avg'] / data['EPS']

# Current Ratio
data['current_ratio'] = data['jumlah_aset_lancar'] / data['jumlah_liabilitas_jangka_pendek']

#Quick Ratio
#data['quick_ratio'] = (data['jumlah_aset_lancar'] - data['persediaan']) / data['jumlah_liabilitas_jangka_pendek']

#Cash Ratio 
data['cash_ratio'] = data['kas'] / data['jumlah_liabilitas_jangka_pendek']

#Account Receivable Turnover
data['AR_Turnover'] = data['pendapatan']/data['piutang']

#Account Receivable Turnover in Days
data['AR_Turnover_in_Days'] = int(365) / data['AR_Turnover']

#Asset Turnover
data['Aset_Turnover'] = data['pendapatan'] / data['jumlah_aset']

#Inventory Turnover
#data['inventory_turnover'] = data['cogs'] / data['persediaan']

#deb to Equity
data['debt_to_equity'] = data['jumlah_liabilitas'] / data['jumlah_ekuitas']

#equity ratio
data['equity_ratio'] = data['jumlah_ekuitas'] / data['jumlah_aset']

#Debt Ratio --> Is it needed? even already had equity ratio?
data['debt_ratio'] = data['jumlah_liabilitas'] / data['jumlah_aset']

#Gross Profit Margin
data['gpm'] = data['laba_bruto'] / data['pendapatan']

#Net Profit Margin
data['net_profit_margin'] = data['laba_rugi_komprehensif'] / data['pendapatan']

#Return on Asset
data['roa'] = data['laba_rugi_komprehensif'] / data['jumlah_aset']

#return on Capital Employed
data['roce'] = data['laba_bruto']/(data['jumlah_ekuitas']+data['jumlah_liabilitas_jangka_panjang'])


data.to_csv('data_olahan.csv', index=False)

# Number of nan

number_of_missing_value = data.isna().sum()

number_of_missing_value



len(data)
